% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coco_dataset-data.R, R/coco_dataset.R
\docType{data}
\name{coco_dataset}
\alias{coco_dataset}
\title{NEK eQTL data.frame for the cis-region.
.}
\format{data.frame}
\usage{
coco_dataset

coco_dataset
}
\description{
NEK eQTL data.frame for the cis-region.
.

NEK eQTL data.frame for the cis-region.
.
}
\examples{
data(coco_dataset)
class(coco_dataset)
dim(coco_dataset)
data(coco_dataset)
class(coco_dataset)
dim(coco_dataset)
}
\keyword{datasets}

