% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coco_ld_matrix-data.R, R/coco_ld_matrix.R
\docType{data}
\name{coco_ld_matrix}
\alias{coco_ld_matrix}
\title{NEK eQTL LD matrix for the cis-region.
.}
\format{Matrix}
\usage{
coco_ld_matrix

coco_ld_matrix
}
\description{
NEK eQTL LD matrix for the cis-region.
.

NEK eQTL LD matrix for the cis-region.
.
}
\examples{
data(coco_ld_matrix)
class(coco_ld_matrix)
dim(coco_ld_matrix)
data(coco_ld_matrix)
class(coco_ld_matrix)
dim(coco_ld_matrix)
}
\keyword{datasets}

