% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coconek1-data.R
\docType{data}
\name{coco_nek1}
\alias{coco_nek1}
\title{NEK eQTL data for the cis-region.
.}
\format{An list of class \code{"coco_dataset"}}
\usage{
coco_nek1
}
\description{
NEK eQTL data for the cis-region.
.
}
\examples{
data(coco_nek1)
class(coco_nek1)
coco_nek1$var_y
dim(coco_nek1$ld_matrix)
}
\keyword{datasets}

