\name{CoCoCore-package}
\alias{CoCoCore-package}
\docType{package}
\title{CoCoCore}
\description{
The 'section' CoCoCore, provides the shared library of CoCo, 
that is the binary of CoCo, 
with the interface function "call.coco" for calling CoCo, 
and some lower level extensions of this one entry point interface function.

The functions of this 'package are not to be called directly by the user. 

% Please use the functions of the package \link{CoCoRaw-package}
% (and \link{CoCoObjects-package}
%  and \link[CoCoGraph]{CoCoGraph-package}). 

% Datasets are found in the packages \link[CoCo]{CoCo-package} and
% \link[CoCoCg]{CoCoCg-package}.
}

% \details{
%    \tabular{ll}{
%       Package: \tab CoCoCore     \cr
%       Type:    \tab Package      \cr
%       Version: \tab 0.1.7.6      \cr
%       Date:    \tab 2010-03-17   \cr
%     Copyright: \tab Copyright (C) by Jens Henrik Badsberg, 1991-2009, 
% 	            non-profit use and redistribution permitted \cr
%    }
% }

\author{
Jens Henrik Badsberg

Maintainer: Jens Henrik Badsberg <coco@badsberg.eu>
}
% \references{ }
\keyword{ package }
