\name{CoCoObjects-package}
\alias{CoCoObjects-package}
\docType{package}
\title{CoCoObjects}
\description{
The 'section' provides classes for objects for CoCo-sessions 
and classes for objects for models in sessions.

This 'section' contains a few functions of interest to the user.

Because of the development of the object system in R and due 
to the development of the CoCo-packages these classes are (was)
for CoCo-objects kept in this separate 'section' (former package).
}

% \details{
%    \tabular{ll}{
%       Package: \tab CoCoObjects  \cr
%       Type:    \tab Package      \cr
%       Version: \tab 0.1.7.6      \cr
%       Date:    \tab 2010-03-17   \cr
%    }
% }

\author{
Jens Henrik Badsberg

Maintainer: Jens Henrik Badsberg <coco@badsberg.eu>
}
% \references{ }
\keyword{ package }
