\name{CoCoRaw-package}
\alias{CoCoRaw-package}
\docType{package}
\title{CoCoRaw}
\description{
  The 'section' provides R functions for the individual 'commands' of CoCo.
% The functions in this package are implemented without objects.
}

% \details{
%    \tabular{ll}{
%       Package: \tab CoCoRaw      \cr
%       Type:    \tab Package      \cr
%       Version: \tab 0.1.7.6      \cr
%       Date:    \tab 2010-03-17   \cr
%    }
% }

\author{
Jens Henrik Badsberg

Maintainer: Jens Henrik Badsberg <coco@badsberg.eu>
}
% \references{ }
\keyword{ package }
