% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_vary.R
\name{estimate_vary}
\alias{estimate_vary}
\title{estimate_vary}
\usage{
estimate_vary(data_set)
}
\arguments{
\item{data_set}{data.frame. The summary statistics data.frame.

Note: Below equation 8 in the cojo paper, is how we estimate the trait variance. \cr
  y'y = D_j S_j^2 (n-1) + D_j B_j^2 \cr
  Trait variance taken as the median of the equation above.}
}
\description{
\code{estimate_vary} Estimate trait variance from summary statistics.
}
\author{
James Boocock and Eli Stahl
}

