% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{get_allowable_correlations}
\alias{get_allowable_correlations}
\title{get_allowable_correlations}
\usage{
get_allowable_correlations(probs, corr)
}
\arguments{
\item{probs}{- marginal probabilities.}
}
\description{
Get allowable correlations
}
\author{
James Boocock
}

