% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint_from_ids.R
\name{joint_from_ids}
\alias{joint_from_ids}
\title{conditional_from_ids}
\usage{
joint_from_ids(rsids, res_preparation, idxs, return_only_these = F,
  joint = T, exact = F)
}
\arguments{
\item{rsids}{character vector - SNPIDs in res preparation to perform CoCo analysis on.}

\item{idxs}{integer vector - postiion in input data to condition on.}

\item{return_only_these}{boolean
\itemize{
\item (T) = return the joint or conditional statistics for only SNPs in idxs or rsids vectors.
\item (F) = return the joint or conditional statistics for SNPs not in the idxs or rsids vectors.
}}

\item{joint}{boolean
\itemize{
 \item(T) = perform joint estimation.
 \item(F) = perform conditional estimation
 }}

\item{res_prepration}{coco_data - the result of running the coco prep_dataset_common command.}
}
\description{
Joint from ids conditional wrapper
}
\details{
\code{joint_from_idxs} Performs a joint or conditional signals from a specific set of SNPs, or indexes.
}
\examples{

data(coco_nek1)
res = joint_from_ids(rsids="rs10520157",coco_nek1,exact=T)
# Second hit rs4417927
res$res_step[res$res_step$rsid == "rs4417927",]

}
\author{
James Boocock and Eli Stahl
}

