% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_betas.R
\name{logistic_to_linear}
\alias{logistic_to_linear}
\title{estimate_betas}
\usage{
logistic_to_linear(b, se, phi, ref_af, mu)
}
\arguments{
\item{data_set}{data.frame. Contains the relevant}

\item{ld_noise}{citation. Pirinen, Matti, Peter Donnelly, and Chris CA Spencer. "Efficient computation with a linear mixed model on large-scale data sets
with applications to genetic studies." The Annals of Applied Statistics 7.1.}
}
\description{
Stepwise conditional wrapper
}
\details{
Generate conditional signals from a conditional dataset.
}
\author{
James Boocock and Eli Stahl
}

