\name{makeCoCo}
\alias{makeCoCo}

\title{ Make a CoCo object }

\description{ Make a CoCo object. }

\usage{
makeCoCo(...)
}

\arguments{
  \item{\dots}{ \code{\link[CoCoCore]{make.coco}}. }
}

\details{ The distributed source code is only able to handle one concurrent
	  CoCo object. Use \code{\link[CoCoRaw]{endCoCo}}. }

% \value{
%   ~Describe the value returned
%   If it is a LIST, use
%   \item{comp1 }{Description of 'comp1'}
%   \item{comp2 }{Description of 'comp2'}
%   ...
% }

\references{ Badsberg, J.H.: Xlisp+CoCo, Aalborg University, 1996. }

\author{ Jens Henrik Badsberg }

% \note{ }

% \section{ToDo}{....}

\seealso{ \code{\link[CoCoCore]{make.coco}}. }

% \examples{ }

\keyword{ interface }% at least one, from doc/KEYWORDS
\keyword{ multivariate }% __ONLY ONE__ keyword per line
