% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_dataset_coco.R
\name{prep_dataset_coco}
\alias{prep_dataset_coco}
\title{estimate_vary}
\usage{
prep_dataset_coco(data_set, ld_matrix, var_y, hwe_variance = F, exact = F,
  use_info = T)
}
\arguments{
\item{data_set}{data.frame. Coco formatted input data.}

\item{ld_matrix}{matrix. LD matrix ordered by postiion.}

\item{var_y}{float. Phenotypic variance for Y, if not specified it will be estimated using \code{\link{estimate_vary}}.}

\item{hwe_variance}{boolean. Force the use of Hardy-weinberg equillibrium when calculating the genotypic variance.}

\item{exact}{boolean. Perform an exact analysis, extra data checks used in \code{\link{prep_data_set_common}}}
}
\description{
Prepare dataset coco.
}
\details{
\code{prep_dataset_coco} Common processing function for checking and parsing a dataset.
}
\examples{
data(coco_dataset)
data(coco_ld_matrix)
coco_data = prep_datset_coco(coco_dataset,coco_ld_matrix)
coco_data$var_y

}
\author{
James Boocock
}

