% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_algebra.R
\name{step_conditional}
\alias{step_conditional}
\title{step_conditional}
\usage{
step_conditional(betas, ld_matrix, neffs, var_y, hwe_diag_outside, hwe_diag,
  return_entire_beta_set = F, exact = F, ses = NULL)
}
\arguments{
\item{betas}{vector of betas that will be adjusted in the matrix operations.}

\item{ld_matrix}{LD matrix for the subset of the region.}

\item{var_y}{variance of of the phenotype}

\item{hwe_diag_outside}{Diagonal matrix containing the genotypic variance}

\item{hwe_diag}{Diagonal matrix containing the genotypic variance *n}

\item{neff}{effective sample size for each SNP in the dataset}
}
\value{
Joint betas and standard errors.
}
\description{
Step conditional
}
\details{
Run a joint step of the analysis.
}
\author{
James Boocock
}

