% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coco_wrappers.R, R/stepwise_run.R
\name{stepwise_coco}
\alias{stepwise_coco}
\title{stepwise_conditional}
\usage{
stepwise_coco(res_preparation, p_value_threshold = 1e-04,
  colinear_threshold = 0.9, joint = T, max_iter = 100,
  return_all_betas = F)

stepwise_coco(res_preparation, p_value_threshold = 1e-04,
  colinear_threshold = 0.9, joint = T, max_iter = 100,
  return_all_betas = F)
}
\arguments{
\item{p_value_threshold}{float P-value threshold for forward-stepwise selection.}

\item{joint}{boolean
\itemize{
 \item(T) = perform joint estimation when doing stepwise selection.
 \item(F) = perform conditional estimation when doing stepwise selection.
 } 
 @export}

\item{data_set}{data.frame. Coco formatted input data.}

\item{ld_matrix}{matrix. LD matrix ordered by postiion}

\item{rsids}{character vector. SNPIDs in res preparation to perform CoCo analysis on.}

\item{idxs}{integer vector - postiion in input data to condition on.}

\item{var_y}{float. Phenotypic variance for Y, if not specified it will be estimated using \code{\link{estimate_vary}}.}

\item{hwe_variance}{boolean. Force the use of Hardy-weinberg equillibrium when calculating the genotypic variance.}

\item{data_set}{input file format}

\item{ld_matrix}{LD matrix for the subset of the region.}

\item{stepwise_results}{- results from a stepwise analysis.}
}
\value{
all_but_ones  betas and standard error for the entire region.
}
\description{
Stepwise coco parameter estimation wrapper.

Forward model selection in coco.
}
\details{
\code{stepwise_coco} wraps \code{\link{joint_from_ids}} for convinience.

\code{stepwise_coco}. Forward model selection in coco.
}
\author{
James Boocock and Eli Stahl

James Boocock
}

