\name{CoCo-internal}
\title{Internal CoCo objects}

% \alias{.First.lib}
% \alias{.onAttach}
% \alias{.onLoad}

\description{Internal CoCo and imported objects.}

\details{These are not to be called by the user.}

\note{
The function \code{dG} was (before bundles was removed by R 2.10.0)
an "alias" for \code{\link[dynamicGraph]{dg}},
but with loading the package "CoCoGraph".
For help on \code{\link[dynamicGraph]{dg}}
use \code{help(dg, package = "dynamicGraph")}.
}

\keyword{internal}
